// tools/conformance/src/modules/nestjs-command-application-generator/nestjs-command-application-generator.types.ts
// 🏷️ Types

import type { Tree } from "@nx/devkit";

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsCommandApplicationGeneratorArguments {
  options: NestjsCommandApplicationGeneratorOptions;
  tree: Tree;
}

/**
 * Auto-generated documentation placeholder.
 */
export interface NestjsCommandApplicationGeneratorOptions {
  destinationRoot?: string;
  name?: string;
}